+++ clients-kt/src/main/kotlin/org/apache/kafka/common/record/DefaultRecordBatch.kt
@@ -330,13 +330,13 @@ class DefaultRecordBatch internal constructor(
         buffer.putInt(PARTITION_LEADER_EPOCH_OFFSET, epoch)
     }
 
-    override fun checksum(): Long = ByteUtils.readUnsignedInt(buffer, CRC_OFFSET)
+    override fun checksum(): UInt = ByteUtils.readUnsignedInt(buffer, CRC_OFFSET)
 
     override val isValid: Boolean
         get() = sizeInBytes() >= RECORD_BATCH_OVERHEAD && checksum() == computeChecksum()
 
-    private fun computeChecksum(): Long =
-        Crc32C.compute(buffer, ATTRIBUTES_OFFSET, buffer.limit() - ATTRIBUTES_OFFSET)
+    private fun computeChecksum(): UInt =
+        Crc32C.compute(buffer, ATTRIBUTES_OFFSET, buffer.limit() - ATTRIBUTES_OFFSET).toUInt()
 
     // note we're not using the second byte of attributes
     private fun attributes(): Byte = buffer.getShort(ATTRIBUTES_OFFSET).toByte()
@@ -487,7 +487,7 @@ class DefaultRecordBatch internal constructor(
 
         override fun lastSequence(): Int = loadBatchHeader().lastSequence()
 
-        override fun checksum(): Long = loadBatchHeader().checksum()
+        override fun checksum(): UInt = loadBatchHeader().checksum()
 
         override fun countOrNull(): Int? = loadBatchHeader().countOrNull()
